refactor(domain): Domain 패키지 정리 및 쓰지 않는 import 정리
