+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,6 +1,6 @@
 package racingcar.domain.car;
 
-import racingcar.strategy.MovingStrategy;
+import racingcar.domain.strategy.MovingStrategy;
 
 public class Car {
