+++ src/main/java/racingcar/domain/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.strategy;
+package racingcar.domain.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
