+++ src/main/java/racingcar/domain/strategy/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.strategy;
+package racingcar.domain.strategy;
 
 import java.util.Random;
