+++ src/test/java/racingcar/domain/NameTest.java
